tlx-ads/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
PyJWT==2.9.0
email-validator==2.2.0

# rate limit redis (multi-instância)
redis==5.0.8

# opentelemetry (observabilidade)
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-requests==0.46b0

# opcional se quiser scheduler via Redis (não obrigatório no MVP)
rq-scheduler==0.13.1
